Remove loggin on alert long poll endpoing
For the moment, only commented out the code in case
we change our minds on this.

The issue is that if there is a client repeating calls
to this ever so often, the log gets filled with these
messages. Considering the value of the message is not
critical, and other approaches are a bit of overkill, it
was decided to simply not log.

Approaches considered include:
  - Log this particular message in separate file
  - Keep a fixed size map of job IDs to last time
  this message was logged and only re-log if a certain
  interval has passed


Former-commit-id: 52b4b608aefad45cf9494cfe20febd03fbe62df5

Original commit: elastic/x-pack@2a0ec7870323742a145748b1e16b6e4c012abbec