allow the system truststore to be used if one is not specified
A truststore should not be required as the default system truststore can be used
to validate certificates that have been signed by most commercial CAs.

Additionally, the HttpClient is now a lifecycle component to prevent out of memory
exceptions when starting up with a bad configuration; when an exception is thrown
in the constructor, Guice will continue to try to create the object until the system runs
out of memory.

Closes elastic/x-pack#476

Original commit: elastic/x-pack@9006590b893cad50a6eb2ec0e07431cf8340f1ba