add support for separate keystore and truststore in HttpClient
The HttpClient class tried to use the truststore as a keystore for watcher with Shield fallback.
This is trappy because in Shield, the default is to only have a keystore that also acts as the
truststore, so the fallback would probably fail in most cases. Additionally, support was missing
for a keystore that had a separate key password.

The new behavior allows for specifying both a keystore and a truststore. If only a keystore is
specified, it will also be used as the truststore. If neither is defined, the system truststore will
be used. All settings fallback to shield settings.

Closes elastic/x-pack#480

Original commit: elastic/x-pack@98c1d0231f30dbb24508c626edb7793e1c1d0da1