allow authentication exceptions for unauthorized anonymous users
Today, we always throw an AuthorizationException for an unauthorized user. This is
problematic when anonymous access is enabled and the HTTP client being used
does not support preemptive basic authentication as only the anonymous user
will be used by such a client.

This change adds a setting to allow an AuthenticationException to be thrown for
anonymous users. This will clients such as browsers to work with anonymous
access and authenticated access.

Closes elastic/x-pack#853

Original commit: elastic/x-pack@f9975e853d4cfd2061ddeb89b01c5a9f99aec416