Replace nulls with empty strings in CSV writer
The input record is fed as input to the transforms.
The input record is populated from each CSV line.
If the line contains empty fields (e.g. "foo,,bar"),
the input record was getting populated with null which
was in return breaking the transforms.

In this commit, the input record is efficiently populated
from the line while securing all null values are replaced
with empty strings.


Former-commit-id: 7e261780a2e1e7bbdc6336e9baf3da298bc1597f

Original commit: elastic/x-pack@980346e3903b7722c154bcaeed70b34d18374595