system actions are only logged at the correct audit level
Today, some system actions could be logged by default when the actions performed are
not internal. Additionally for internal actions, we never checked if the user was the system
user.

This adds a check to ensure the user is the System user and that the actions that are being
suppressed are known system actions.

Closes elastic/x-pack#902

Original commit: elastic/x-pack@502e6a2a98c7bce5dfa98768345fa2b49a1628ef