store names of indices as an array instead of a string
We currently store the names of indices as a comma separated string instead
of an array. An array is the proper format for this information so this commit
changes the index audit trail to store the indices as an array.

Closes elastic/x-pack#917

Original commit: elastic/x-pack@e54fe3855b28ac01b574ebe5ed1ade4d35a1c869