rename shield audit indices to keep naming consistent
This change renames the shield audit indices to keep naming consistent with other plugins.
The name of the index uses '_' to separate words, a '-' to separate the prefix from the time
portion, and '.'s to separate the different portions of the date.

Closes elastic/x-pack#925

Original commit: elastic/x-pack@7df38c2c4127efb2c510cb75cf41cdf756419443