only load the shield plugin for remote auditing client
Currently, any plugin that is on the classpath will be loaded by the TransportClient we create
for remote audit indexing. The only plugin that should be loaded for this client is the Shield
plugin. To accomplish this, we disable loading of plugins from the classpath and specify that
the Shield plugin should be loaded.

Closes elastic/x-pack#941

Original commit: elastic/x-pack@de624824d0c90f5db1b63eed07cb7127a0d92543