Attach zip assembly as artifact to build when deploying internally
Attaching the zip assembly as an artifact to the build will include it when deploying to maven repositories. This will in turn facilitate testing Elasticsearch snapshot versions in a Found environment. We need to to that in order to cut down the delay from a release is made until it is available on Found.

Original commit: elastic/x-pack@8eef4072e7061b8463561aa8f2824f254b8a9f03