Bug 695 - Store both function name and the functionDescription
This changes the function field of AnomalyRecord and AnomalyCause
to contain the function as inserted in the detector. However,
it also adds a new field, functionDescription, where the user
friendly name is stored for display or other purposes.

This change was also necessary in order to implement bug-1136
as we need to use the actual function name to pick the correct
leaf normaliser.


Former-commit-id: 389a5affb78343cb99e3010baca54b45b3b4d011

Original commit: elastic/x-pack@fa822177563a55a0825477d2735a4890851dece0