[docs] change `active-directory` to `active_directory`
In the documentation reorgantization for Shield, the type of the active directory realm changed
and is not valid. Users are configuring Shield incorrectly and running into issues because of
this. This changes the realm type to the correct value.

Closes elastic/x-pack#291

Original commit: elastic/x-pack@5ada6725b825ae1478d4ece3c90d4ec0b2181957