don't attach audit user to remote requests and fix rest origin address
Currently, we attach the internal audit user to all requests. This is incorrect for requests that
need to be sent to a remote cluster. For these cases, we should require a user to be defined
to access the remote cluster if it is protected by Shield.

Additionally, the origin_address field for rest request fields is formatted differently than other
address fields. This changes the field to only be the remote address.

Closes elastic/x-pack#278
Closes elastic/x-pack#279

Original commit: elastic/x-pack@d98531a111ef2814b569892eb8d602d3c4409192