use NetworkUtils instead of InetAddress.getLocalHost
Prior to this commit, we were InetAddress.getLocalHost() to get the hostname and host
address when auditing. This is different than how we report the node's hostname and host
address in other places where we use NetworkUtils. This caused false failures to be seen
with the IndexAuditTrail tests. This commit switches the audit trails to use the NetworkUtils
methods.

Closes elastic/x-pack#285

Original commit: elastic/x-pack@c596aab71138f00f54325160b5381306241786c3