Add mechanism for transporting shard-level actions by node
Currently, many shard-level operations are transported with a request
per shard via TransportBroadcastAction. This results in an inefficiency
on large clusters. A more efficient mechanism would be to aggregate the
shard-level requests and transport them in a request per node.

This commit introduces TransportNodeBroadcastAction which is the
high-level mechanism for transporting the shard-level operations in a
single request per node. The shard-level operations are still executed
asynchronously (in parallel subject to thread pool availability) on the
receiving node and per-node shard-level results are aggregated into a
single response per node. These node-level results are then aggregated
into a single response to the initial request.

Closes #7990