Remove cycle between job.config.verification and job.transform
The cycle exists because the JobConfigurationVerifier method
checkTransfomOutputIsUSed is throwing TransformConfigurationException.

However, the issue that we try to detect there is not a transform
consistency problem but rather a job config problem where some
of the transforms specified are not being used. Therefore, we
can easily solve this by throwing a JobConfigurationException instead.


Former-commit-id: bfd49d3148af605fdcac393c9c4b089d116e2462

Original commit: elastic/x-pack@6dbc84f7310fb3befbf3eebca880db80b747b9a9